+++ django/ft_transcendence/chat/chat_utils.py
@@ -1,0 +1,15 @@
+from chat.models import Chat
+from django.db.models import Q
+
+def get_or_create_chat(request, chatWithUser):
+	refUser = request.user
+
+	chat = Chat.objects.filter(
+			Q(fromUser=refUser, toUser=chatWithUser) |
+			Q(fromUser=chatWithUser, toUser=refUser)
+		).first()
+
+	if not chat:
+		chat = Chat.objects.create(fromUser=request.user, toUser=chatWithUser)
+
+	return chat
