+++ django/ft_transcendence/chat/static/chat/css/room.css
@@ -1,94 +1,97 @@
 body {
-	font-family: Arial, sans-serif;
-	background-color: #f4f4f9;
-	margin: 0;
-	padding: 0;
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-	justify-content: center;
-	height: 100vh;
+    font-family: Arial, sans-serif;
+    background-color: #f4f4f9;
+    margin: 0;
+    padding: 0;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+}
+
+.chat-wrapper {
+    display: flex;
+    height: 60vh;
+    width: 65vw;
+    background-color: white;
+    border-radius: 8px;
+    overflow: hidden;
+    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
+}
+
+.chat-history-container {
+    flex: 0.3;
+    background-color: #f0f0f0;
+    padding: 10px;
+    overflow-y: auto;
+}
+
+.chat-bar {
+    display: flex;
+    flex-direction: column;
+    gap: 15px;
 }
 
 .chat-container {
-	width: 90%;
-	max-width: 1200px;
-	background-color: white;
-	border-radius: 8px;
-	box-shadow: 0 4px 10px rgba(0, 0, 0, 0.1);
-	padding: 20px;
+    flex: 0.7;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
 }
 
 .chat-log {
-	width: 100%;
-	height: 400px;
-	border: 1px solid #ddd;
-	padding: 10px;
-	margin-bottom: 15px;
-	border-radius: 5px;
-	overflow-y: scroll;
-	background-color: #fafafa;
-}
-
-.message {
-	padding: 10px;
-	margin: 5px;
-	border-radius: 8px;
-	width: fit-content;
-	max-width: 70%;
-	word-wrap: break-word;
-	clear: both;
-}
-
-.message-sender {
-	background-color: #d1e7dd; /* Light green for sender */
-	align-self: flex-end;
-	text-align: right;
-	margin-left: auto;
-}
-
-.message-receiver {
-	background-color: #f8d7da; /* Light red for receiver */
-	align-self: flex-start;
-	text-align: left;
-	margin-right: auto;
+    width: 100%;
+    height: 100%;
+    padding: 10px;
+    overflow-y: scroll;
+    background-color: #fafafa;
+    border: 1px solid #ddd;
+    border-radius: 5px;
 }
 
 .chat-input {
-	display: flex;
-	justify-content: space-between;
-	margin-top: 15px;
+    display: flex;
+    gap: 10px;
 }
 
 #chat-message-input {
-	width: 80%;
-	padding: 10px;
-	border-radius: 5px;
-	border: 1px solid #ddd;
-	font-size: 16px;
+    flex: 1;
+    padding: 10px;
+    border-radius: 5px;
+    border: 1px solid #ddd;
 }
 
 #chat-message-submit {
-	width: 18%;
-	padding: 10px;
-	background-color: #28a745;
-	color: white;
-	font-size: 16px;
-	border: none;
-	border-radius: 5px;
-	cursor: pointer;
+    width: 100px;
+    background-color: #28a745;
+    color: white;
+    border: none;
+    border-radius: 5px;
+    cursor: pointer;
 }
 
 #chat-message-submit:hover {
-	background-color: #218838;
+    background-color: #218838;
 }
 
-.chat-header {
-	text-align: center;
-	margin-bottom: 10px;
+.message {
+    padding: 10px;
+    margin: 5px;
+    border-radius: 8px;
+    max-width: 70%;
+    clear: both;
+    word-wrap: break-word;
 }
 
-.chat-header h1 {
-	font-size: 1.5em;
-	color: #333;
+.message-sender {
+    background-color: #d1e7dd;
+    text-align: right;
+    margin-left: auto;
+}
+
+.message-receiver {
+    background-color: #f8d7da;
+    text-align: left;
+    margin-right: auto;
 }
