+++ django/ft_transcendence/chat/templates/chat/direct_message.html
@@ -8,12 +8,31 @@
 		<meta charset="UTF-8">
 		<meta name="viewport" content="width=device-width, initial-scale=1.0">
 		<link rel="stylesheet" href="{% static 'css/main.css' %}">
-		<link rel="stylesheet" href="{% static 'css/direct_message.css' %}">
-		<title>Profile Page</title>
+		<link rel="stylesheet" href="{% static 'chat/css/direct_message.css' %}">
+		<link rel="stylesheet" href="{% static 'core/css/social.css' %}">
+
+		<title>Direct Message</title>
 	</head>
 
-	<body>
+	<body class="scroll">
+		<div class="chat-container">
+			<!---------- Background Video ---------->
+			<video autoplay muted loop class="video-background">
+				<source src="{% static 'core/images/planet-background.mp4' %}" type="video/mp4">
+				Your browser does not support the video tag.
+			</video>
 
+			<div class="chat-bar">
+			{% for chat in user_chats %}
+				<div class="social--user-profile-avatar">
+					<img src="{{ chat.toUser.userprofile.avatar.url }}" class="social--user-avatar-img">
+				</div>
+				<div class="chat--history">
+					<p class="test">{{ chat.toUser }}  {{ chat.last_message }} </p>
+				</div>
+			{% endfor %}
+			</div>
+		</div>
 	</body>
 
 </html>
