+++ django/ft_transcendence/chat/views.py
@@ -5,7 +5,7 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 import json
 from django.utils.safestring import mark_safe
-from core.utils.profile_utils import get_or_create_chat
+from .chat_utils import get_or_create_chat
 from django.shortcuts import redirect
 
 @login_required
@@ -17,12 +17,14 @@ def room(request, room_name, userID):
     from_user = request.user.id
     to_user   = User.objects.get(id=userID)
     user_profile = User.objects.get(id=userID)
+    user_chats = Chat.get_user_chats(request.user)
 
     context = {
         "room_name": room_name,
         "userID": to_user.id,
         "sender_username": request.user.username,
-        "receiver_username": to_user.username
+        "receiver_username": to_user.username,
+        "user_chats": user_chats
     }
     return render(request, "chat/room.html", context)
 
@@ -36,4 +38,12 @@ def get_room_redirect(request, userID):
 # ---- Direct Message -------------------
 @login_required
 def direct_message(request):
-    return render(request, "chat/direct_message.html")
+    user_chats = Chat.get_user_chats(request.user)
+    last_user_chat = user_chats.order_by('createdAt').first().toUser.id
+
+    # context     = {
+    #     "user_chats": user_chats
+    # }
+
+    return get_room_redirect(request, userID=last_user_chat)
+    # return render(request, "chat/direct_message.html", context)
