+++ django/ft_transcendence/core/utils/profile_utils.py
@@ -2,21 +2,6 @@ from django.contrib.auth.decorators import login_required
 from chat.models import Chat
 from django.db.models import Q
 
-
-@login_required
-def get_or_create_chat(request, chatWithUser):
-	refUser = request.user
-
-	chat = Chat.objects.filter(
-			Q(fromUser=refUser, toUser=chatWithUser) |
-			Q(fromUser=chatWithUser, toUser=refUser)
-		).first()
-
-	if not chat:
-		chat = Chat.objects.create(fromUser=request.user, toUser=chatWithUser)
-
-	return chat
-
 def _has_friend_request_(request, user_profile):
 	friend_request_receiver = request.user.receiver.all()
 	friend_request_sender   = request.user.sender.all()
